New python notebook added
The law of larger number supports the intuition that as the size of the sample increases, the mean value of the sample will better approximate the mean or expected value in the population. As the sample size goes to infinity,  the sample mean will converge to the population mean. This project is a demonstration of this assumption.